+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Ride file creation from lookup
 - Added reference contributor for functions in one file
 - Added find usages for functions in one file
+- Added completion for user-defined functions
 
 ### Fixed
 - Language grammar errors